hotfix(flag_friends): fix the bug and the split function. And i refactor parser args
