+++ generate_mirror_workflow.py
@@ -1,0 +1,26 @@
+#!/usr/bin/env python3
+##
+## @VyOkPROJECT, 2023
+## mirror-generator
+## File description:
+## why do you look at this aweful code ...
+##
+
+import subprocess
+
+def generate_mirror_workflow(binary_name: str, repo_ssh: str, repo_name:str):
+    mirror_name = f"{repo_name}-mirror"
+    f = open("verif_mirror.yml", "r")
+    mirror_file_data = f.read()
+    f.close()
+    mirror_file_data = mirror_file_data.replace("MIRROR_URL_MIRROR_GENERATOR", repo_ssh)
+    mirror_file_data = mirror_file_data.replace("EXECUTABLE_MIRROR_GENERATOR", binary_name)
+    actions = [(["mkdir", f"{binary_name}/{mirror_name}/.github"], "Folder already exist: " + f"{binary_name}/{mirror_name}/.github"),
+               (["mkdir", f"{binary_name}/{mirror_name}/.github/workflows"], "Folder already exist: " + f"{binary_name}/{mirror_name}/.github/workflows")]
+    for action in actions:
+        if subprocess.run(action[0]).returncode == 1:
+            print(action[1])
+            exit(84)
+    mirror_file = open(f"{binary_name}/{mirror_name}/.github/workflows/verif_mirror.yml", "w")
+    mirror_file.write(mirror_file_data)
+    mirror_file.close()
